Add files via upload
This is an etl script for a credit transaction dataset. In this script, the data was extracted from kaggle using KaggleAPI. The transformation process includes: removing duplicates, assigning appropriate data types to columns, extracting values from columns and handling null values. For loading, the loading is to a mssql database. The credential should be in an os file and the script would pick it from there. The mode of authentication is windows authentication, hence, the absence of username and password. Finally, the loading process loads the records in a bit size of 50 records.